requirements.txt
aiogram>=3.0,<4.0
aiohttp>=3.9,<4.0
python-dotenv>=1.0,<2.0
